refactor(box-shape): refactor and add unit tests
